Rename namespace from Compile to Compiler across multiple files
