+++ src-deprecated/Compiler/LazyModule.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace BEAR\Package;
+namespace BEAR\Package\Compiler;
 
 use BEAR\AppMeta\AbstractAppMeta;
+use BEAR\Package\Module;
 use BEAR\Package\Module\ResourceObjectModule;
 use BEAR\Package\Module\ScriptInjectorModule;
 use Ray\Compiler\LazyModuleInterface;
