+++ src-deprecated/Compiler/NewInstance.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace BEAR\Package\Compile;
+namespace BEAR\Package\Compiler;
 
 use Ray\Di\Exception\Unbound;
 use Ray\Di\InjectorInterface;
